move get_project_name() to helpers
