+++ sdks/python/src/opik/api_objects/helpers.py
@@ -1,10 +1,18 @@
 import datetime
+import logging
+
 import uuid_extensions
 
 from typing import List, Any
 
 from .. import datetime_helpers
 from typing import Optional
+
+from ..config import OPIK_PROJECT_DEFAULT_NAME
+from ..logging_messages import NESTED_SPAN_PROJECT_NAME_MISMATCH_WARNING_MESSAGE
+
+
+LOGGER = logging.getLogger(__name__)
 
 
 def generate_id() -> str:
@@ -22,3 +30,35 @@ def datetime_to_iso8601_if_not_None(
 
 def list_to_batches(items: List[Any], batch_size: int) -> List[List[Any]]:
     return [items[i : i + batch_size] for i in range(0, len(items), batch_size)]
+
+
+def get_project_name(
+    parent_project_name: Optional[str],
+    child_project_name: Optional[str],
+    show_warning: bool = True,
+) -> Optional[str]:
+    if parent_project_name != child_project_name:
+        # if the user has specified a project name -> print warning
+        if show_warning and child_project_name is not None:
+            # if project name is None -> use default project name
+            parent_project_name_msg = (
+                parent_project_name
+                if parent_project_name is not None
+                else OPIK_PROJECT_DEFAULT_NAME
+            )
+            child_project_name_msg = (
+                child_project_name
+                if child_project_name is not None
+                else OPIK_PROJECT_DEFAULT_NAME
+            )
+
+            LOGGER.warning(
+                NESTED_SPAN_PROJECT_NAME_MISMATCH_WARNING_MESSAGE.format(
+                    child_project_name_msg, parent_project_name_msg
+                )
+            )
+        project_name = parent_project_name
+    else:
+        project_name = child_project_name
+
+    return project_name
